GLZA v0.5
GLZA v0.5 has a lot of changes vs. v0.4, mostly targeted at improving
compression ratios but also at improving memory use. The most significant
changes I remember off the top of my head are:

GLZAformat:

Improved delta encoding. Instead of just subtracting bytes for strides of 1, 2
or 4, it now detects strides of 1 - 4 and for even strides attempts to
determine whether the data is big-endian or little endian and for strides of 4
determine whether the data has one 4 byte channel or two 2 byte channels. Added
reordering of data blocks by "columns" for files with strides of 2 and 4. It is
not perfect, but helps compression significantly on GEO, mr, x-ray, .wav files,
.bmp files, .tif files, etc.

Reworked and lowered the threshold for classification of data as text. Added
capital lock encoding (can be disabled with -l0) and automatic space after
linefeed additions for capital encoded files.

GLZAcompress:

The suffix tree nodes now have a length field and the code uses this to build
and use a more normal suffix tree (with branches containing more than 1
symbol). Now it is normal except if the non-overlapping count changes there
will be an extra branch/node for each non-overlapping count. The node sibling
array size was changed from 4 to 2 to reduce the node size. These changes could
have been used to reduce memory usage but for large files instead the savings
(by default) is used to read more data. For the typical corpus files, the now
usually fit in memory right from the start but it can take quite a while to
build and evaluate the suffix tree. Memory usage can be set with the -r#
option, with # being millions of bytes (was -m# and it worked differently).

The default production cost is no longer fixed at 6.0 bits. Now at the start of
each duduplication cycle it is set to 1.5 plus log2(number of symbols/alphabet
size).

For capital encoded files, the first deduplication pass only analyzes strings
that start with a space and end before a space. This can be disabled with -w0.

GLZAencode: Removed a couple of inefficiencies, tweaked some models and
improved memory use.

GLZAdecode: Moved capital decoding to thread #2. Added string length fields to
the production rule structure to support a hierarchical dictionary strings and
reuse of dictionary memory. Also added code to reuse expired production rule
structures. Memory use is much better.

Overall compression is generally a little better (and a lot better on certain
files) but occassionally worse. Compression can be significantly slower,
especially at the start. Decompression is typically about 10% slower.
Decompression RAM use is quite good. I see enwik9 decompression RAM of less
than 200 MB of physical memory (for default parameters) vs. 330 MB for v0.4
(which was already Pareto Frontier).